requirements.txt
# TODO: pin versions to ranges
jira
GitPython==0.1.7
python-dateutil
argh
argcomplete
